更新方法应用
